refactor permissions into nu_utils
