+++ crates/nu-utils/src/lib.rs
@@ -2,6 +2,7 @@ mod casing;
 pub mod ctrl_c;
 mod deansi;
 pub mod emoji;
+pub mod filesystem;
 pub mod locale;
 pub mod utils;
 
@@ -16,3 +17,6 @@ pub use deansi::{
     strip_ansi_likely, strip_ansi_string_likely, strip_ansi_string_unlikely, strip_ansi_unlikely,
 };
 pub use emoji::contains_emoji;
+
+#[cfg(unix)]
+pub use filesystem::users;
